+++ projects/ng-porcelain/src/lib/combobox/combobox/combobox.component.ts
@@ -22,7 +22,10 @@ type ItemType = string | object;
 @Component({
 	selector: 'porcelain-combobox, p-combobox',
 	templateUrl: './combobox.component.html',
-	styleUrls: ['./combobox.component.scss']
+	styleUrls: ['./combobox.component.scss'],
+	host: {
+		'[class.combobox--has-focus]': 'hasFocus'
+	}
 })
 export class ComboboxComponent extends Loggable implements OnInit {
 	/**