+++ projects/ng-porcelain/src/lib/refiners/refiners.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { DateRefinerModule } from '../date-refiner/date-refiner.module';
 import { SimpleRefinerModule } from '../simple-refiner/simple-refiner.module';
+import { SearchRefinerModule } from '../search-refiner/search-refiner.module';
 
 import { RefinersComponent } from './refiners/refiners.component';
 
@@ -11,7 +12,8 @@ export const REFINERS_DIRECTIVES = [RefinersComponent];
 export const REFINERS_IMPORTS = [
 	CommonModule,
 	DateRefinerModule,
-	SimpleRefinerModule
+	SimpleRefinerModule,
+	SearchRefinerModule
 ];
 
 @NgModule({