+++ projects/ng-porcelain/src/lib/refiners/refiners/refiners.component.ts
@@ -1,12 +1,16 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { BaseRefinerDefinition } from '../../shared/types/Refiners/BaseRefinerDefinition';
+import { Loggable } from '../../Loggable';
+import { isEqual } from 'lodash-es';
 
 @Component({
 	selector: 'porcelain-refiners, p-refiners',
 	templateUrl: './refiners.component.html',
 	styleUrls: ['./refiners.component.scss']
 })
-export class RefinersComponent implements OnInit {
+export class RefinersComponent extends Loggable implements OnInit {
+	readonly name = 'RefinersComponent';
+
 	// Inputs
 	@Input() refiners: BaseRefinerDefinition[];
 	@Input() allowIncompleteEmit: boolean = true;
@@ -19,14 +23,19 @@ export class RefinersComponent implements OnInit {
 	// State
 	values: { [slug: string]: string[] } = {};
 
-	constructor() {}
+	constructor() {
+		super();
+	}
 
 	ngOnInit() {}
 
 	handleRefinerChange(update: [string, any]) {
 		let [slug, selected] = update;
 
-		this.setValue(slug, selected);
+		if (!isEqual(this.values[slug], selected)) {
+			this.debug('handleRefinerChange(update)', { before: this.values[slug], after: selected });
+			this.setValue(slug, selected);
+		}
 	}
 
 	setValue(slug: string, value: any) {
@@ -40,5 +49,9 @@ export class RefinersComponent implements OnInit {
 
 	isDateRefiner(refiner: BaseRefinerDefinition): boolean {
 		return refiner.type === 'date';
+	}
+
+	isSearchRefiner(refiner: BaseRefinerDefinition): boolean {
+		return refiner.type === 'search';
 	}
 }