+++ projects/ng-porcelain/src/lib/shared/types/Refiners/RefinerType.ts
@@ -1,4 +1,1 @@
-export enum RefinerType {
-	simple = 'simple',
-	date = 'date'
-}
+export type RefinerType = 'simple' | 'date' | 'search';